backend/requirements.txt
fastapi[standard]
langchain
langchain-core
langchain-community>=0.3.23
langchain-openai
langchain-text-splitters
pydantic<2
langgraph
duckduckgo-search>=4.1.1
typing-extensions>=4.7.1
faiss-cpu
googlesearch-python
bs4
httpx
openai
tiktoken